Refactor Value to Test conversion from Evaluation
